post-share: move connections warning into primary component
